fonction cupidon

+++ jeux.h
@@ -5,10 +5,11 @@
 #include <time.h>
 #include "interface.h"
 #include "fonction.h"
+#include <assert.h>
 
 void attribution(Role* tabRoles, unsigned short nbrDeJoueurs);
 int voteCapitaine(GUI* input);
 int voteFinDeTour(GUI* input, Role* tabRoles, unsigned short nbrDeJoueursEnVie, int caseCapitaine);
 void fVoyante(GUI* input, Role* tabRoles);
 int fLoupgarou(GUI* input, Role* tabRoles, unsigned short nbrDeJoueursEnVie);
-void fCupidon(GUI* input, Role* tabRoles, int tabCupidon[]);
+void fCupidon(GUI* input, Role* tabRoles, unsigned short tabCupidon[]);
